Upgrade to code-gen v1.10. (#48)
However, there were some manual changes required to the fake clientset generated code. The newly generated
code also included a call to object.Watch() but there was no corresponding implementation for it. This code
was manually deleted. It should be safe since this was in the fake clientset code, the build succeeds and
the unit tests complete successfully.

Testing Done:

1. make test passes
2. Built and deployed the sensor-controller. Ran the example sensor. It completed successfully.

Refs #46