pubsub: separate client sometimes unavoidable
If the topic is in another project and doesn't already exist, we need a separate client to create the topic. There's no `Client#CreateTopicInProject` or `Topic#Create` function. This being the case, we may as well go ahead and create a separate client whenever the topic project ID is different from the main project ID.