Allow watching of resources in all namespaces at once (#334)
The `namespace` field of a resource event source is now optional to
allow for configuration that would watch for resource related events in
all namespaces at once. Simply omitting the field works in this way
because the k8s dynamic client used in implementation can accept a
zero-length string for namespace.