[git-sensor-artifact] Force fetch from git repository
We've run into an issue with using the existing git sensor with our repository.  The sensor works for a while until it starts erroring on `failed to fetch. err: some refs were not updated`.  This seems to happen when we merge code to our repository, and our deploy process updates tags.  Updating tags on existing branches causes git to throw an error unless you force fetch.  The artifact already force fetches when given a particular ref, and I can't think of a downside of doing it everywhere given that you wouldn't be making local changes to branches in a sensor.

Signed-off-by: rory.lamendola <rory.lamendola@petalcard.com>