Reset kafka consumer on cleanup
When a partition is unassigned from one consumer and assigned to
a different consumer, we need to clear the trigger state. If we
don't clear this state false trigger invocations can occur if
a partition is reassigned back to a consumer that it was previously
assigned to because stale messages will still be in memory.

Signed-off-by: David Farr <david_farr@intuit.com>