feat(bitbucketserver): multiple improvements to bitbucket server event source
This Pull Request contains 3 improvements to the Bitbucket Server Event source.

1. Customizable Apply Webhook Check Interval

Currently the apply webhook check runs every 60 seconds which could be to often. This is now configurable.

2. Manage Webhooks for Entire Projects

When a list of Bitbucket Server Projects are provided the event source will ensure a webhook exists in all repositories in the provided project.

3. Ability to Skip Refs Changed when a Pull Request is Open

A new option has been added that adds a check when a wehook is received, when a repo:refs_changed event is received a check will be ran to skip publishing the event if a Pull Request is opened. This is to get around the issue that Bitbucket will send this event even when a Pull Request is opened and there is already a pr:from_ref_updated event that would launch multiple workflow. This allows us to have a branch, pull request and tag workflows.

Signed-off-by: Ryan Currah <ryan@currah.ca>